`SerializablePoint` now uses `u32` in order to safely implement `Hash` and `Eq`

+++ src/message.rs
@@ -57,13 +57,13 @@ pub enum SelectionType {
 pub enum ElementProperty {
     Text(String),
     ShiftText(String),
-    TextPositionX(f32),
-    TextPositionY(f32),
+    TextPositionX(u32),
+    TextPositionY(u32),
     FollowCaps,
     Boundary(usize, Option<SerializablePoint>),
     Keycode(usize, Option<u32>),
-    MouseSpeedIndicatorPositionX(f32),
-    MouseSpeedIndicatorPositionY(f32),
+    MouseSpeedIndicatorPositionX(u32),
+    MouseSpeedIndicatorPositionY(u32),
     MouseSpeedIndicatorRadius(f32),
 }
